pyproject.toml
[tool.poetry]
name = "adonai"
version = "0.0.1"
description = "Lightweight identification, authentication and authorization service"
authors = ["Aleksander Lavrov <egnod@ya.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
sitri = "^0.7.0"
graphene-sqlalchemy = "^2.2.2"
flask-migrate = "^2.5.2"
psycopg2-binary = "^2.8.4"
passlib = "^1.7.2"
argon2-cffi = "^19.2.0"
flask-shell-ipython = "^0.4.1"
flask-graphql = "^2.0.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
pre-commit = "^1.21.0"
flask-jwt = "^0.3.2"
python-dotenv = "^0.10.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
